refactor: consume MappingRuleMatcher and OidcPrincipalLoader from CSL
Switch all OC consumers of the two AuthN utilities to the CSL packages and
delete the OC copies. Part of #244 (parent #95).

- AuthenticationHandler: imports io.camunda.security.core.oidc.OidcPrincipalLoader
- 7 MappingRuleMatcher consumers (zeebe/engine x5, search-domain, service):
  imports io.camunda.security.core.auth.MappingRuleMatcher
- Delete security/security-core/.../auth/MappingRuleMatcher.java,
  OidcPrincipalLoader.java, and their tests
- Drop now-unused json-path and jcip-annotations deps from
  security/security-core/pom.xml
- Add camunda-security-library-core dependency to search/search-domain,
  service, and zeebe/engine where it was missing

The CSL classes are behaviour-compatible with the OC originals, with two
parallel improvements landed earlier in CSL OidcPrincipalLoader and now in
the migrated MappingRuleMatcher: claim names are sanitised into JSONPath
expressions (so plain names match rather than silently fail), and JSONPath
evaluation failures are logged at DEBUG with claim keys only (no values, no
PII leak).

The CSL pin is temporarily set to 0.1.0-SNAPSHOT so the build resolves
against camunda/camunda-security-library#353. The TODO on
parent/pom.xml#L48 must be replaced with the released CSL alpha version
that contains MappingRuleMatcher before this PR can leave draft.
