+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/processing/identity/authorization/AuthorizationCheckBehavior.java
@@ -10,8 +10,8 @@ package io.camunda.zeebe.engine.processing.identity.authorization;
 import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
 import com.google.common.cache.LoadingCache;
-import io.camunda.security.auth.MappingRuleMatcher;
 import io.camunda.security.configuration.EngineSecurityConfig;
+import io.camunda.security.core.auth.MappingRuleMatcher;
 import io.camunda.zeebe.engine.EngineConfiguration;
 import io.camunda.zeebe.engine.processing.Rejection;
 import io.camunda.zeebe.engine.processing.identity.AuthorizedTenants;
